chore(#9): update folder structure
