+++ src/main/java/com/zoltraak/gateway/adapters/gpu/GpuProviderPort.java
@@ -1,7 +1,7 @@
 package com.zoltraak.gateway.adapters.gpu;
 
 import com.zoltraak.gateway.domain.enums.PodStatus;
-import com.zoltraak.gateway.domain.shared.PodConnectionDetails;
+import com.zoltraak.gateway.domain.models.provider.PodConnectionDetails;
 import reactor.core.publisher.Mono;
 
 public interface GpuProviderPort {
