+++ src/main/java/com/zoltraak/gateway/features/gpu/GpuController.java
@@ -2,7 +2,7 @@ package com.zoltraak.gateway.features.gpu;
 
 import com.zoltraak.gateway.adapters.gpu.GpuProviderPort;
 import com.zoltraak.gateway.domain.enums.PodStatus;
-import com.zoltraak.gateway.domain.shared.GatewayResponse;
+import com.zoltraak.gateway.domain.response.GatewayResponse;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
